backend/requirements.txt
# Lizard backend
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Document parsing
pymupdf==1.25.1
python-pptx==1.0.2

# OpenAI-compatible LLM
openai==1.55.3

# RAG: Redis + embeddings
redis>=5.0.0
numpy>=1.24.0

# CORS
starlette==0.41.3
